refactor: Game, Board components and some utils, add draw functionality, change some exports to default exports
