+++ src/pages/game/types.ts
@@ -1,0 +1,3 @@
+import { PlayerSymbol } from "@/resources/players";
+
+export type SquareSymbol = PlayerSymbol | null;
