+++ src/pages/game/utils/calculateIndex.ts
@@ -1,0 +1,7 @@
+export function calculateIndex(
+	columnIndex: number,
+	rowIndex: number,
+	size: number,
+) {
+	return columnIndex * size + rowIndex;
+}
