+++ src/pages/game/utils/checkDraw.ts
@@ -1,14 +1,9 @@
-import { PlayerSymbol } from "@/resources/players";
+import { SquareSymbol } from "../types";
 
-function checkWhetherAllSquaresAreFilled(
-	squares: (PlayerSymbol | null)[],
-): boolean {
+function checkWhetherAllSquaresAreFilled(squares: SquareSymbol[]): boolean {
 	return squares.every((square) => square !== null);
 }
 
-export function checkDraw(
-	isWinner: boolean,
-	squares: (PlayerSymbol | null)[],
-): boolean {
+export function checkDraw(isWinner: boolean, squares: SquareSymbol[]): boolean {
 	return !isWinner && checkWhetherAllSquaresAreFilled(squares);
 }
