+++ src/pages/game/utils/symbolMap.ts
@@ -1,6 +1,7 @@
 import { PlayerSymbol } from "@/resources/players";
+import { SquareSymbol } from "../types";
 
-export const SYMBOL_MAP = new Map<PlayerSymbol | null, string | null>([
+export const SYMBOL_MAP = new Map<SquareSymbol, string | null>([
 	[null, null],
 	[PlayerSymbol.Cross, "X"],
 	[PlayerSymbol.Circle, "O"],
