+++ src/resources/players/types.ts
@@ -1,6 +1,7 @@
 import { PlayerSymbol } from "./enums";
 
 export type Player = {
+	id: string;
 	name: string;
 	score: number;
 	symbol: PlayerSymbol;
