+++ src/store/playerSlice/selectors.ts
@@ -1,3 +1,4 @@
+import { createSelector } from "@reduxjs/toolkit";
 import { RootState } from "../store";
 
-export const selectPlayers = (state: RootState) => state.player.players
+export const selectPlayers = (state: RootState) => state.player.players;
